Add pretty-print functions to sast
Note that the pretty print functions in sast.ml are not tested,
although they can compile. It would be easy to identify the bugs upon
testing, so I didn't handwrite sast's myself to test them. The next step
would be writing type checking functions, like in microc/semant.ml.
These will be tricky, but not too hard. After getting the hello-world
devliverable done, we should add A LOT OF missing features to the
frontend, before we proceed to code generation. We need also consult the
workflow with Yunlan.