Check function arguments
Currently the arguments to function calls can be checked, both in terms
of numbers and types. But there's one bug. If we have
FUNC(a be int)
{
	b = a;
}
our compiler will say this "a is undeclared", because we didn't add
FUNC#a to f_sym_table. This will be fixed in next commit.