irgen: expr: Use ref to store local_vars
local_vars in MicroC is predetermined because MicroC uses a C90-style
function structure: local variables must be declared before statements.
Hence MicroC doesn't need to update local_vars, which we obviously need
to. This simply changes local_var to a ref so that we can update
whenever accessing it. In the future, we should put all accesses to
local_var in specific functions instead of in the lengthy pattern
matching.