irgen: Finish basic GCD support
A hacky implementation in scanner: I have changed all the functions from
upper case to lower case, so that when user types MAIN, it will be main
and can be interpreted by lli directly as the program entry point.

This tells us all of our internal view of functions are lower case.
Built-in functions should be the same, such as PRINT.